F OpenNebula/one#7077: NUMA PCIe topology
This commits adds support for generating a NUMA PCIe hierarchy by adding
a pcie-expander-bus in each virtual node. When a PCI passthrough or
SR-IOV device is added is attached to the same virtual node as the
physical server. Changes include:

 - Generate VM_* device parameters when the allocation is performed
 - Generate CONTEXT PCIi_ADDRESS parameters at deployment time
 - Refactor of the HostSharePCI to simplify argument passing.

The hierarchy created is as follows (example 2 NUMA nodes):

                      +-----------+
                      |           |
                      | PCIe Root |
                      |           |
                      +--+-----+--+
                         | ... |
                   +-----+     +-----+
    NUMA node 0    |                 |    NUMA node 1
              +----+-----+      +----+-----+
              | expander |      | expander |
              |    bus   |      |    bus   |
              +----+-----+      +----+-----+
                   |                 |
                   |                 ...
                   |
                   +------> port (0)
                   |        ...
                   +------> port (3)
                   |
                   |    +-----------+
                   | up |   PCIe    +-----> downstream port (0)
                   +----+  Switch   |       ...
                        |           +-----> downstream port (8)
                        +-----------+

+++ src/host/HostShareNUMA.cc
@@ -1052,8 +1052,8 @@ bool HostShareNUMA::schedule_nodes(NUMANodeRequest &nr, unsigned int threads,
 int HostShareNUMA::make_hugepage_topology(HostShareCapacity &sr,
                                           unsigned long hpsz_kb, bool do_alloc)
 {
-    unsigned long n_hp = sr.mem / hpsz_kb; //sr.mem = SUM(sr.nodes.memory)
-    unsigned long node_fhp = 0;
+    long long n_hp = sr.mem / hpsz_kb; //sr.mem = SUM(sr.nodes.memory)
+    long long node_fhp = 0;
 
     int node_id = -1;
 
@@ -1156,7 +1156,7 @@ int HostShareNUMA::make_affined_topology(HostShareCapacity &sr, int node_id,
     // -------------------------------------------------------------------------
     // Check that the node has enough free huge pages
     // -------------------------------------------------------------------------
-    unsigned long n_hp = 0;
+    long long n_hp = 0;
 
     HostShareNode::HugePage * hpage = nullptr;
